# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * PhotonLint API — Design Rule Checker for Silicon Photonics
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Per-layer design rules for a specific GDS process layer
 */
export interface FoundryLayer {
  /** GDS layer number */
  layer: number;
  /** GDS datatype number */
  datatype: number;
  /** Human-readable layer name (e.g. "Si Strip Waveguide") */
  name: string;
  /** Minimum feature width in microns */
  minWidth: number;
  /** Minimum spacing between features in microns */
  minSpacing: number;
  /** Minimum bend radius in microns (0 = not applicable for this layer) */
  minBendRadius: number;
}

/**
 * Custom PDK rule values to override the bundled YAML defaults
 */
export interface FoundryOverrideInput {
  /** Manufacturing grid snap size in microns */
  gridSize: number;
  /** Per-layer rule overrides */
  layers: FoundryLayer[];
}

/**
 * YAML default values — present only when hasOverride is true, for diff display
 */
export type FoundryDefaults = {
  gridSize: number;
  layers: FoundryLayer[];
};

/**
 * A foundry Process Design Kit with per-layer design rules
 */
export interface Foundry {
  id: string;
  name: string;
  description: string;
  /** Technology node / process name */
  technology: string;
  /** Manufacturing grid snap size in microns (applies to all layers) */
  gridSize: number;
  /** Per-layer design rules — each configured GDS layer has its own thresholds */
  layers: FoundryLayer[];
  /** True when engineer-supplied custom rules are active for this foundry */
  hasOverride: boolean;
  /** YAML default values — present only when hasOverride is true, for diff display */
  defaults?: FoundryDefaults;
}

/**
 * A single polygon in the GDS layout
 */
export interface LayoutPolygon {
  layer: number;
  datatype: number;
  /** Polygon vertices as [[x, y], ...] in microns */
  vertices: number[][];
  /** True if this polygon's layer is configured in the selected PDK */
  isPdkLayer?: boolean;
}

/**
 * A PDK-configured layer that was checked
 */
export interface ConfiguredLayer {
  layer: number;
  datatype: number;
  name: string;
}

/**
 * Bounding box of the entire layout in microns
 */
export interface LayoutBounds {
  minX: number;
  minY: number;
  maxX: number;
  maxY: number;
}

/**
 * Full layout geometry for visual rendering
 */
export interface LayoutData {
  bounds: LayoutBounds;
  polygons: LayoutPolygon[];
  /** List of PDK-configured layers that were DRC-checked */
  configuredLayers?: ConfiguredLayer[];
}

export type DrcViolationSeverity = typeof DrcViolationSeverity[keyof typeof DrcViolationSeverity];


export const DrcViolationSeverity = {
  critical: 'critical',
  warning: 'warning',
  info: 'info',
} as const;

/**
 * A single DRC rule violation found in the layout
 */
export interface DrcViolation {
  /** Name of the violated design rule */
  rule: string;
  /** The required value per the PDK */
  requirement: string;
  /** Location of the violation in the layout */
  location: string;
  severity: DrcViolationSeverity;
  /** Additional details about the violation */
  details: string;
  /**
     * Polygon vertices [[x, y], ...] in microns for visual highlighting
     * @nullable
     */
  geometry?: number[][] | null;
}

export type DrcResultStatus = typeof DrcResultStatus[keyof typeof DrcResultStatus];


export const DrcResultStatus = {
  pass: 'pass',
  fail: 'fail',
  error: 'error',
} as const;

/**
 * Complete DRC result for a layout file
 */
export interface DrcResult {
  id: string;
  foundryId: string;
  foundryName: string;
  filename: string;
  status: DrcResultStatus;
  violationCount: number;
  passedChecks: number;
  totalChecks: number;
  violations: DrcViolation[];
  /** @nullable */
  errorMessage: string | null;
  checkedAt: string;
  processingTimeMs: number;
  layoutData?: LayoutData | null;
}

export type DrcRunSummaryStatus = typeof DrcRunSummaryStatus[keyof typeof DrcRunSummaryStatus];


export const DrcRunSummaryStatus = {
  pass: 'pass',
  fail: 'fail',
  error: 'error',
} as const;

/**
 * Summary of a past DRC run
 */
export interface DrcRunSummary {
  id: string;
  foundryId: string;
  foundryName: string;
  filename: string;
  status: DrcRunSummaryStatus;
  violationCount: number;
  checkedAt: string;
  processingTimeMs: number;
}

/**
 * A frequently occurring violation type
 */
export interface CommonViolation {
  rule: string;
  count: number;
  percentage: number;
}

/**
 * Aggregate statistics across all DRC runs
 */
export interface DrcStats {
  totalRuns: number;
  passCount: number;
  failCount: number;
  errorCount: number;
  /** Percentage of runs that passed (0-100) */
  passRate: number;
  totalViolations: number;
  avgViolationsPerRun: number;
  commonViolations: CommonViolation[];
}

export interface ApiError {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListDrcRunsParams = {
limit?: number;
};

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};