# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PhotonLint API — Design Rule Checker for Silicon Photonics
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /foundries/{id}/override:
    put:
      operationId: setFoundryOverride
      tags: [foundries]
      summary: Save custom PDK rule values for a foundry
      description: |
        Upserts engineer-supplied rule values for the given foundry. When an override
        is active, the DRC engine uses these values instead of the bundled YAML defaults.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoundryOverrideInput"
      responses:
        "200":
          description: Updated foundry with the applied override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Foundry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Foundry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteFoundryOverride
      tags: [foundries]
      summary: Remove custom PDK rule values and revert to defaults
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Foundry reverted to YAML defaults
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Foundry"
        "404":
          description: Foundry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
...
  schemas:
...
    FoundryLayer:
      type: object
      description: Per-layer design rules for a specific GDS process layer
      properties:
        layer:
          type: number
          description: GDS layer number
        datatype:
          type: number
          description: GDS datatype number
        name:
          type: string
          description: Human-readable layer name (e.g. "Si Strip Waveguide")
        minWidth:
          type: number
          description: Minimum feature width in microns
        minSpacing:
          type: number
          description: Minimum spacing between features in microns
        minBendRadius:
          type: number
          description: Minimum bend radius in microns (0 = not applicable for this layer)
      required: [layer, datatype, name, minWidth, minSpacing, minBendRadius]
    FoundryOverrideInput:
      type: object
      description: Custom PDK rule values to override the bundled YAML defaults
      properties:
        gridSize:
          type: number
          description: Manufacturing grid snap size in microns
        layers:
          type: array
          description: Per-layer rule overrides
          items:
            $ref: "#/components/schemas/FoundryLayer"
      required: [gridSize, layers]
    Foundry:
      type: object
      description: A foundry Process Design Kit with per-layer design rules
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        technology:
          type: string
          description: Technology node / process name
        gridSize:
          type: number
          description: Manufacturing grid snap size in microns (applies to all layers)
        layers:
          type: array
          description: Per-layer design rules — each configured GDS layer has its own thresholds
          items:
            $ref: "#/components/schemas/FoundryLayer"
        hasOverride:
          type: boolean
          description: True when engineer-supplied custom rules are active for this foundry
        defaults:
          type: object
          description: YAML default values — present only when hasOverride is true, for diff display
          properties:
            gridSize:
              type: number
            layers:
              type: array
              items:
                $ref: "#/components/schemas/FoundryLayer"
          required: [gridSize, layers]
      required: [id, name, description, technology, gridSize, layers, hasOverride]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...