# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * PhotonLint API — Design Rule Checker for Silicon Photonics
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
}